+++ src/activities/ActivityResult.h
@@ -31,8 +31,8 @@ struct PercentResult {
   int percent = 0;
 };
 
-struct AutoPageTurnResult {
-  uint32_t seconds = 0;
+struct IntervalResult {
+  uint32_t value = 0;
 };
 
 struct PageResult {
@@ -68,7 +68,7 @@ struct FilePathResult {
 };
 
 using ResultVariant = std::variant<std::monostate, WifiResult, KeyboardResult, MenuResult, ChapterResult, PercentResult,
-                                   AutoPageTurnResult, PageResult, SyncResult, NetworkModeResult, FootnoteResult,
+                                   IntervalResult, PageResult, SyncResult, NetworkModeResult, FootnoteResult,
                                    BookmarkResult, FileBrowserActionResult, FilePathResult>;
 
 struct ActivityResult {